chore: refactor and improve redis keys, update projects.example.json contracts
