+++ pooler/callback_modules/pair_total_reserves.py
@@ -27,14 +27,10 @@ from pooler.utils.models.message_models import PowerloomCallbackProcessMessage
 from pooler.utils.models.message_models import UniswapPairTotalReservesSnapshot
 from pooler.utils.redis.rate_limiter import load_rate_limiter_scripts
 from pooler.utils.redis.redis_keys import (
-    uniswap_cb_broadcast_processing_logs_zset,
+    cb_broadcast_processing_logs_zset,
 )
-from pooler.utils.redis.redis_keys import (
-    uniswap_discarded_query_pair_total_reserves_epochs_redis_q_f,
-)
-from pooler.utils.redis.redis_keys import (
-    uniswap_failed_query_pair_total_reserves_epochs_redis_q_f,
-)
+from pooler.utils.redis.redis_keys import discarded_query_epochs_redis_q
+from pooler.utils.redis.redis_keys import failed_query_epochs_redis_q
 
 
 class PairTotalReservesProcessor(CallbackAsyncWorker):
@@ -212,7 +208,6 @@ class PairTotalReservesProcessor(CallbackAsyncWorker):
                         epoch_broadcast,
                     )
                     queued_epochs.append(epoch_broadcast)
-                # uniswap_failed_query_pair_total_reserves_epochs_redis_q_f.format(current_epoch.contract)
                 failed_query_epochs = await self._redis_conn.lpop(
                     failed_query_epochs_key,
                 )
@@ -266,11 +261,11 @@ class PairTotalReservesProcessor(CallbackAsyncWorker):
     ):
         # check for enqueued failed query epochs
         epochs = await self._prepare_epochs(
-            failed_query_epochs_key=uniswap_failed_query_pair_total_reserves_epochs_redis_q_f.format(
-                msg_obj.contract,
+            failed_query_epochs_key=failed_query_epochs_redis_q.format(
+                self._stream, msg_obj.contract,
             ),
-            discarded_query_epochs_key=uniswap_discarded_query_pair_total_reserves_epochs_redis_q_f.format(
-                msg_obj.contract,
+            discarded_query_epochs_key=discarded_query_epochs_redis_q.format(
+                self._stream, msg_obj.contract,
             ),
             current_epoch=msg_obj,
             snapshot_name='pair reserves',
@@ -283,8 +278,8 @@ class PairTotalReservesProcessor(CallbackAsyncWorker):
             cb_fn_async=self._fetch_token_reserves_on_chain,
             enqueue_on_failure=enqueue_on_failure,
             data_source_contract_address=msg_obj.contract,
-            failed_query_redis_key=uniswap_failed_query_pair_total_reserves_epochs_redis_q_f.format(
-                msg_obj.contract,
+            failed_query_epochs_key=failed_query_epochs_redis_q.format(
+                self._stream, msg_obj.contract,
             ),
             transformation_lambdas=[],
         )
@@ -296,7 +291,7 @@ class PairTotalReservesProcessor(CallbackAsyncWorker):
         cb_fn_async,
         enqueue_on_failure,
         data_source_contract_address,
-        failed_query_redis_key,
+        failed_query_epochs_key,
         transformation_lambdas: List[Callable],
         **cb_kwargs,
     ):
@@ -332,7 +327,7 @@ class PairTotalReservesProcessor(CallbackAsyncWorker):
                     contract=data_source_contract_address,
                 )
                 await self._redis_conn.rpush(
-                    failed_query_redis_key,
+                    failed_query_epochs_key,
                     queue_msg_obj.json(),
                 )
                 self._logger.debug(
@@ -369,7 +364,7 @@ class PairTotalReservesProcessor(CallbackAsyncWorker):
         self, broadcast_id, update_state,
     ):
         await self._redis_conn.hset(
-            uniswap_cb_broadcast_processing_logs_zset.format(self.name),
+            cb_broadcast_processing_logs_zset.format(self.name),
             broadcast_id,
             json.dumps(update_state),
         )
