+++ pooler/static/projects.example.json
@@ -13,21 +13,21 @@
         "0x3041cbd36888becc7bbcbc0045e3b1f144466f5f",
         "0x9c4fe5ffd9a9fc5678cfbd93aa2d4fd684b67c4c",
         "0xd3d2e2692501a5c9ca623199d38826e513033a17",
-        "0x149148acc3b06b8cc73af3a10e84189243a35925",
+        "0x9fae36a18ef8ac2b43186ade5e2b07403dc742b1",
         "0xbb2b8038a1640196fbe3e38816f3e67cba72d940",
         "0xa478c2975ab1ea89e8196811f51a7b7ade33eb11",
         "0x11181bd3baf5ce2a478e98361985d42625de35d1",
         "0x7b73644935b8e68019ac6356c40661e1bc315860",
         "0x0af81cd5d9c124b4859d65697a4cd10ee223746a",
-        "0x252b9a9ad1e85d9eff8a3e4782491f87e7109e3a",
+        "0xab400c46c830a2f87939dcfdcbfaaadf76f35721",
         "0x470e8de2ebaef52014a47cb5e6af86884947f08c",
-        "0x582e3da39948c6339433008703211ad2c13eb2ac"
-        ],
+        "0xf89ce5ed65737da8440411544b0499c9fad323b2"
+      ],
       "workers": [{
         "module": "pooler.callback_modules.pair_total_reserves",
         "class_name": "PairTotalReservesProcessor",
         "name": "PowerLoom|UniswapPooler|Callbacks|PairTotalReservesProcessor",
-        "num_instances": 8
+        "num_instances": 6
       }]
     },
     {
@@ -44,21 +44,21 @@
         "0x3041cbd36888becc7bbcbc0045e3b1f144466f5f",
         "0x9c4fe5ffd9a9fc5678cfbd93aa2d4fd684b67c4c",
         "0xd3d2e2692501a5c9ca623199d38826e513033a17",
-        "0x149148acc3b06b8cc73af3a10e84189243a35925",
+        "0x9fae36a18ef8ac2b43186ade5e2b07403dc742b1",
         "0xbb2b8038a1640196fbe3e38816f3e67cba72d940",
         "0xa478c2975ab1ea89e8196811f51a7b7ade33eb11",
         "0x11181bd3baf5ce2a478e98361985d42625de35d1",
         "0x7b73644935b8e68019ac6356c40661e1bc315860",
         "0x0af81cd5d9c124b4859d65697a4cd10ee223746a",
-        "0x252b9a9ad1e85d9eff8a3e4782491f87e7109e3a",
+        "0xab400c46c830a2f87939dcfdcbfaaadf76f35721",
         "0x470e8de2ebaef52014a47cb5e6af86884947f08c",
-        "0x582e3da39948c6339433008703211ad2c13eb2ac"
-        ],
+        "0xf89ce5ed65737da8440411544b0499c9fad323b2"
+      ],
       "workers": [{
         "module": "pooler.callback_modules.trade_volume",
         "class_name": "TradeVolumeProcessor",
         "name": "PowerLoom|UniswapPooler|Callbacks|TradeVolumeProcessor",
-        "num_instances": 8
+        "num_instances": 6
       }]
     }
   ]
